Prevent hook loops when restarting the server.
Restarting the server doesn't reload jfx_bridge, so we'd grab the previously hooked local_call and treat that as the real local_call, leading to hooks calling hooks and sadness. Now, we save the real local_call to a new name only if it hasn't already been saved, and then place our hook - this means that we always have the original local_call, and the new hooks won't call old hooks by mistake.