Improve local_call/local_eval hooks
Cleaned up the code a lot using the new jfx_bridge hooks that have been added.
Also added the ability for local_call to inspect the callable to see if it's IDA-related code
and only use execute_sync if it is. This a) prevents the blocking caused by calling IDA-related code in a local __init__ for a class that inherits from an IDA class, and b) avoids running code on the main IDA thread that doesn't need it.