Fix for calling Python 2 IDA APIs with strings
jfx_bridge automatically converts any string into unicode when it goes to a jfx_bridge server running on Python 2. However, IDAPython's APIs expect a plain string, so this will throw an error when calling any API which uses a string. Not a problem on Python3 because it doesn't have the unicode string type.

Fixed in hook_local_call(), by casting any unicode arguments/kwargs that come from the jfx_bridge to plain strings. Not fixed in hook_local_eval at this time, as can't tell if you're calling an IDA API, or might legitimately want a unicode string passed in as an argument. If you're running into issues there, cast the argument that's getting converted into unicode back to a str in the eval string. For example, instead of:
    i.remote_eval("idaapi.set_cmt(ea, comment, False)", ea=ea, comment=comment)
try:
    i.remote_eval("idaapi.set_cmt(ea, str(comment), False)", ea=ea, comment=comment)